stylistic / cosmetic changes in BaseData

+++ processing/base_data.py
@@ -37,7 +37,7 @@ class BaseData(object):
 		featureNames may be None if the object is to have default names, or a list or dict defining the
 		featureName mapping.
 
- 		"""
+		"""
 		super(BaseData, self).__init__()
 		self._nextDefaultValue = 0
 		self._setAllDefault()
@@ -288,9 +288,9 @@ class BaseData(object):
 		raise NotImplementedError
 
 		#MR to find how many of each value
-		def mapperCount (point):
+		def mapperCount(point):
 			return [(point[index],1)]
-		def reducerCount (identifier, values):
+		def reducerCount(identifier, values):
 			total = 0
 			for value in values:
 				total += value
@@ -328,7 +328,7 @@ class BaseData(object):
 		random.seed(seed)
 		if extractionProbability is None:
 			raise ArgumentException("Must provide a extractionProbability")
-		if  extractionProbability <= 0:
+		if extractionProbability <= 0:
 			raise ArgumentException("extractionProbability must be greater than zero")
 		if extractionProbability >= 1:
 			raise ArgumentException("extractionProbability must be less than one")
@@ -357,7 +357,7 @@ class BaseData(object):
 
 		"""
 		# note: we want truncation here
-		numInFold = self.points() / numFolds
+		numInFold = int(self.points() / numFolds)
 		if numInFold == 0:
 			raise ArgumentException("Must specifiy few enough folds so there is a point in each")
 
@@ -742,7 +742,7 @@ class BaseData(object):
 			featureNameNum = int(featureName[len(DEFAULT_PREFIX):])
 			self._nextDefaultValue = max(self._nextDefaultValue + 1, featureNameNum)
 
-		features  = len(self.featureNamesInverse)
+		features = len(self.featureNamesInverse)
 		self.featureNamesInverse[features] = featureName
 		self.featureNames[featureName] = features
